Read in integers and determine the median
Goal of the program is to read in integers and determine the median value.

Program is designed to large data inputs, such that it will read in an integer and determine where to place it with in the list using binary search. This is much more efficient than inserting then sorting.